ray: Add a module to represent rays within a scene
This commit introduces a ray module that defines a Ray type.

A Ray is defined by a point of origin and a direction vector. It
includes an at(time) function that evaluates the ray equation
P(t) = Origin + t * Direction to find the Vec3 position along the ray at
a given time t.